+++ app/controllers/foothold/pages_controller.rb
@@ -1,0 +1,8 @@
+module Foothold
+  class PagesController < ApplicationController
+    def show
+      @page = @site.pages.find(params[:id])
+      @findings = @page.findings.open.order(:check)
+    end
+  end
+end